+++ CHANGELOG.md
@@ -12,8 +12,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add excel function EXACT(value1, value2) support - [595](https://github.com/PHPOffice/PhpSpreadsheet/pull/595)
 - Support workbook view attributes for Xlsx format - [#523](https://github.com/PHPOffice/PhpSpreadsheet/issues/523)
 - Read and write hyperlink for drawing image - [#490](https://github.com/PHPOffice/PhpSpreadsheet/pull/490)
-- Fix ISFORMULA() function to work with a cell reference to another worksheet
-- Added calculation engine support for the new functions that were added in MS Excel 2013 and MS Excel 2016
+- Added calculation engine support for the new bitwise functions that were added in MS Excel 2013
+  - BITAND()     Returns a Bitwise 'And' of two numbers
+  - BITOR()      Returns a Bitwise 'Or' of two number
+  - BITXOR()     Returns a Bitwise 'Exclusive Or' of two numbers
+  - BITLSHIFT()  Returns a number shifted left by a specified number of bits
+  - BITRSHIFT()  Returns a number shifted right by a specified number of bits
+- Added calculation engine support for other new functions that were added in MS Excel 2013 and MS Excel 2016
   - Text Functions
     - CONCAT()        Synonym for CONCATENATE()
     - NUMBERVALUE()   Converts text to a number, in a locale-independent way
@@ -44,6 +49,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fix ISFORMULA() function to work with a cell reference to another worksheet
 - Xlsx reader crashed when reading a file with workbook protection - [#553](https://github.com/PHPOffice/PhpSpreadsheet/pull/553)
 - Cell formats with escaped spaces were causing incorrect date formatting - [#557](https://github.com/PHPOffice/PhpSpreadsheet/issues/557)
 - Could not open CSV file containing HTML fragment - [#564](https://github.com/PHPOffice/PhpSpreadsheet/issues/564)